+++ src/types.ts
@@ -1,3 +1,5 @@
+/** HoloHash Types */
+
 /**
  * @public
  */
@@ -30,6 +32,27 @@ export type AnyDhtHash = HoloHash;
  * @public
  */
 export type ExternalHash = HoloHash;
+/**
+ * @public
+ */
+export type DhtOpHash = HoloHash;
+/**
+ * @public
+ */
+export type WarrantHash = HoloHash;
+/**
+ * @public
+ */
+export enum HoloHashType {
+  Agent = "agent",
+  Entry = "entry",
+  DhtOp = "dhtop",
+  Warrant = "warrant",
+  Dna = "dna",
+  Action = "action",
+  Wasm = "wasm",
+  External = "external",
+}
 
 /**
  * @public
@@ -70,6 +93,18 @@ export type ActionHashB64 = HoloHashB64;
  * @public
  */
 export type AnyDhtHashB64 = HoloHashB64;
+/**
+ * @public
+ */
+export type ExternalHashB64 = HoloHashB64;
+/**
+ * @public
+ */
+export type DhtOpHashB64 = HoloHashB64;
+/**
+ * @public
+ */
+export type WarrantHashB64 = HoloHashB64;
 
 /**
  * @public