+++ CHANGELOG.md
@@ -5,9 +5,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## \[Unreleased\]
 
 ### Added
+- Added alias utility classes for ergonomic use of the new utility classes outlined below, keyed to specific types: `AgentPubKeyB64Map`, `DnaHashB64Map`, `WasmHashB64Map`, `EntryHashB64Map`, `ActionHashB64Map`, `AnyDhtHashB64Map`, `ExternalHashB64Map`, `DhtOpHashB64Map`, `WarrantHashB64Map`, `AgentPubKeyMap`, `DnaHashMap`, `WasmHashMap`, `EntryHashMap`, `ActionHashMap`, `AnyDhtHashMap`, `ExternalHashMap`, `DhtOpHashMap`, `WarrantHashMap`, `LazyAgentPubKeyMap`, `LazyDnaHashMap`, `LazyWasmHashMap`, `LazyEntryHashMap`, `LazyActionHashMap`, `LazyAnyDhtHashMap`, `LazyExternalHashMap`, `LazyDhtOpHashMap`, `LazyWarrantHashMap`. [#392](https://github.com/holochain/holochain-client-js/pull/392)
+- Added utility classes `HoloHashB64Map`, `HoloHashMap`, `MemoizedHoloHashMap`, `DnaHoloHashMap` which provide ergonomic Maps for common use cases where data is keyed by `HoloHash`. [#392](https://github.com/holochain/holochain-client-js/pull/392)
+- Added utility functions `getHashType`, `hashFromContentAndType` for getting `HoloHashType` for a hash, and for generating a hash from a `HoloHashType`. [#392](https://github.com/holochain/holochain-client-js/pull/392)
+- Added enum `HoloHashType` for referencing a HoloHash type by name. [#392](https://github.com/holochain/holochain-client-js/pull/392)
+- Added missing type aliases `WarrantHash`, `WarrantHashB64`, `ExternalHashB64`. [#392](https://github.com/holochain/holochain-client-js/pull/392)
+- Added missing types `DetailsType`, `Details`, `EntryDhtStatus`, `EntryDetails`, `RecordDetails`, `ValidationStatus`, `ValidationReceipt`. [#388](https://github.com/holochain/holochain-client-js/pull/388)
+- Added missing type aliases `DhtOpHash` and `DhtOpHashB64`. [#388](https://github.com/holochain/holochain-client-js/pull/388)
+- Added type alias `DnaDefinition`, which is deprecated and will be removed in 0.22.0. [#388](https://github.com/holochain/holochain-client-js/pull/388)
+- Exported types `PeerUrl`, `SpaceId`. [#391](https://github.com/holochain/holochain-client-js/pull/391)
 ### Fixed
 ### Changed
+- Renamed `DnaDefinition` to `DnaDef`. [#388](https://github.com/holochain/holochain-client-js/pull/388)
 - Update response type of `AdminWebsocket.dumpNetworkStats` to `ApiTransportStats` [#383](https://github.com/holochain/holochain-client-js/pull/383)
+- Update all dependencies to their latest version. [#390](https://github.com/holochain/holochain-client-js/pull/390)
 ### Removed
 
 ## 2025-11-14: v0.20.0